feat:Make some error correction

+++ src/utils/LocalStorage.jsx
@@ -244,8 +244,16 @@ const admin = [
   },
 ];
 export const setLocalStorage = () => {
-  localStorage.setItem("employees", JSON.stringify(employees));
-  localStorage.setItem("admin", JSON.stringify(admin));
+  const existingEmployees = localStorage.getItem("employees");
+  const existingAdmin = localStorage.getItem("admin");
+
+  if (!existingEmployees) {
+    localStorage.setItem("employees", JSON.stringify(employees));
+  }
+
+  if (!existingAdmin) {
+    localStorage.setItem("admin", JSON.stringify(admin));
+  }
 };
 export const getLocalStorage = () => {
   const employees = JSON.parse(localStorage.getItem("employees"));
